add upgrade tasks to main
